WIP: Implement System Environment Extension
This commit starts an implementation of the System Environment Extension. The missing implementation is disallowing it's usage if the test is parallel, as tests would mess with each other's Environments.

This commit also fixes a small issue with SystemProperties extension, in which they wouldn't accept a suspended function. A test for that behavior was added.

This commit also renames the package, to stop stacktrace filtering, so that the error is clear whenever one of these tests fail.

Related to #623